Fix some compiler warnings.
git-svn-id: file:///tmp/tmp.jyKeACR6iO/svn-version/trunk@24 f84e74b7-95bb-4e84-9b5c-02b1165c2552

+++ directory.c
@@ -134,7 +134,7 @@ int afp_rename_request(struct afp_volume *volume,
 
 }
 
-int afp_createdir_request(struct afp_volume * volume, unsigned int dirid, const char * pathname, unsigned short *did_p)
+int afp_createdir_request(struct afp_volume * volume, unsigned int dirid, const char * pathname, unsigned int *did_p)
 {
 	struct {
 		struct dsi_header dsi_header __attribute__((__packed__));
